# Improve performance of sample! function