@Tortar 2023-03-08
I'm running the test for a sample of 1 million and the previous implementation hasn't returned anything yer, probably a 1000x at least :D